+++ index.d.ts
@@ -1,56 +1,86 @@
-export = PageSystem;
-declare class PageSystem {
-    constructor(client: Discord.Client);
-    client: Discord.Client;
+import {
+    Client,
+    InteractionCollector,
+    UserResolvable,
+    MessageEmbed,
+    Interaction,
+    Message,
+    MessageOptions,
+    MessageEditOptions,
+    MessageActionRow,
+    MessageButtonOptions,
+    MessageButtonStyle,
+    ButtonInteraction
+} from "discord.js";
+
+type Context = Interaction | Message
+
+export = Dispage;
+declare class Dispage {
+    constructor(client: Client);
+    client: Client;
     index: number;
-    id: string;
-    embeds: Discord.MessageEmbed[];
-    message: any;
-    interaction: any;
-    reply: Discord.Message;
-    collector: Discord.InteractionCollector<Discord.ButtonInteraction<Discord.CacheType>>;
+    embeds: MessageEmbed[];
+    message: Message;
+    interaction: Interaction;
+    collector: InteractionCollector<ButtonInteraction>;
+    reply: Message;
     ended: boolean;
     started: boolean;
     deleted: boolean;
     duration: number;
     mainStyle: string;
-    buttons: Discord.MessageButtonOptions[];
-    footer: (index: any, total: any) => string;
-    filter: (i: any) => boolean;
+    buttons: {
+        customId: string;
+        emoji: string;
+        style: string;
+        label: string;
+    }[];
+    footer: (index: number, total: number) => string;
+    filter: (int: ButtonInteraction) => boolean;
+    showDisabledButtons: boolean;
+    private _userIDs: Set<string>;
+    get users(): string[];
     get ctx(): "MESSAGE" | "INTERACTION";
-    get type(): "message" | "interaction";
+    get type(): string;
     get endUntill(): number;
-    get currentEmbed(): Discord.MessageEmbed;
-    addButton(o: Discord.MessageButtonOptions): this;
+    get currentEmbed(): MessageEmbed;
+    _getId(user: UserResolvable): string;
+    setMainStyle(style: MessageButtonStyle): this;
+    shouldShowDisabledButtons(should?: boolean): this;
+    removeUser(user: UserResolvable): this;
+    addUser(user: UserResolvable): this;
+    setUser(user: UserResolvable): this;
+    /** @deprecated Use .setUser() instead. */
+    setUserID(user: UserResolvable): this;
+    addButton(o: object): this;
     removeButton(customId: string): this;
-    editButton(customId: string, o: Discord.MessageButtonOptions): this;
-    getRow(disabled?: boolean): Discord.MessageActionRow | null;
-    setUserID(user: Discord.UserResolvable): this;
+    editButton(customId: string, o: object): this;
+    getRows(disabled?: boolean): MessageActionRow | null;
     setFooter(func: Function): this;
-    setEmbeds(embeds: Discord.MessageEmbed[]): this;
-    addEmbed(embed: Discord.MessageEmbed): this;
-    fixEmbeds(embeds: Discord.MessageEmbed | Discord.MessageEmbed[] | Discord.MessageEmbed[][]): Discord.MessageEmbed[];
-    addEmbeds(embeds: Discord.MessageEmbed[]): this;
+    setEmbeds(embeds: MessageEmbed[]): this;
+    addEmbed(embed: MessageEmbed): this;
+    addEmbeds(embeds: MessageEmbed[]): this;
+    _fixEmbeds(embeds: MessageEmbed | MessageEmbed[] | MessageEmbed[][]): MessageEmbed[];
     setDuration(duration: number): this;
     addDuration(duration: number): this;
-    next(): Promise<void> | import(".");
-    previous(): Promise<void> | import(".");
+    next(): Promise<void> | this;
+    previous(): Promise<void> | this;
     setIndex(index: number): this;
-    fixEmbedFooters(): void;
-    edit(opts: Discord.MessageEditOptions): Promise<Discord.Message<boolean>>;
-    disableButtons(): Promise<Discord.Message<boolean>>;
-    end(): Promise<void> | Promise<Discord.Message<boolean>>;
-    delete(): Promise<Discord.Message<boolean>>;
-    update(): Promise<Discord.Message<boolean>>;
+    _fixEmbedFooters(): void;
+    edit(opts: MessageEditOptions): Promise<Message<boolean>>;
+    disableButtons(): Promise<Message<boolean>>;
+    end(): Promise<void> | Promise<Message<boolean>>;
+    delete(): Promise<Message<boolean>>;
+    update(): Promise<Message<boolean>>;
     getOpts(disabled?: boolean): {
-        embeds: Discord.MessageEmbed[];
-        components: Discord.MessageActionRow[];
+        embeds: MessageEmbed[];
+        components: MessageActionRow;
     };
     isMessage(): boolean;
     isInteraction(): boolean;
     canEdit(): boolean;
-    isValidCtx(ctx: any): boolean;
-    checkForErrors(msg: any): string[];
-    start(ctx: any): Promise<import(".")>;
+    isValidCtx(ctx: Context): boolean;
+    checkForErrors(msg: Context): string[];
+    start(ctx: Context): this;
 }
-import Discord = require("discord.js");